Removal of "Maintenance" Portion of the Proposal
The "Maintenance" section of 1017 is unenforceable by code. Future generations of network users cannot be held to binding provisions set forth by the current generation of users unless agreed upon by those future users, and only by code. The decision to remove the section from the Proposal, as opposed to simply moving it to an appendix for reference, is to maintain focus of the proposal and to reduce the potential for confusion by readers, users, and developers.