Reject ECIP-1051 "Ethereum Classic Treasury system".
There has been no work on this proposal (or the earlier IOHK treasury proposal).
There is broad consensus that changing the block rewards to add treasury model is unacceptable.
The Monetary Policy (ECIP-1017) defines the ETC system and gives multi-decade certainty to all ETC ecosystem participants.

If we need to have a discussion of this on ETC Core Developers call then we should.
And then Reject it.