Reject ECIP-1057 "Cold Staking"
Other than Dexaran himself, there is no support for this proposal with the ETC ecosystem.
Proposals of this kind (like PoolTogether on ETH - https://www.pooltogether.com/) can happen just fine with no need for consensus across the ecosystem.

I propose we Reject this proposal.   It could be passed as Informational, I suppose, but there is really no need for there to be an ECIP for this application level proposal.

Even at application level, you would not need/want an ECIP for this until there were multiple implementations and this was become an architectural best practice.

Best, I think, to reject.    A new ECBP could be written which generalized the pattern and that might be accepted.