Reject ECIP-1052 "Smart-contract Security Auditing core"
This requires the Treasury.
Both the Treasury and this Auditing proposal on top run counter to ECIP-1017 monetary policy.

Auditing is a good idea, but should not be centralized into the block rewards.

This runs counter to ETC philosophy and principles and should be rejected.

Perhaps this is raised in an ETC Core Developers call, and formally rejected?