_specs/ecip-1100.md: use current.time offset instead of proposed.time
Using the local head timestamp instead of the proposed
head timestamp better translates the opinion that
first-to-come chains are 'better' than later-arriving
ones. Thus is better to use as a domain for antigravity.

Further, use of local timestamps avoids questions
and variables about the case where an antagonist would
be willing to work extra hard to create blocks with
lower or manipulated timestamps in order to change
the parameter of the curve's imposition.

Signed-off-by: meows <b5c6@protonmail.com>